+++ pyvista/core/filters.py
@@ -34,7 +34,7 @@ import pyvista
 from pyvista.utilities import (FieldAssociation, NORMALS, assert_empty_kwargs,
                                generate_plane, get_array, vtk_id_list_to_array,
                                wrap, ProgressMonitor)
-from pyvista.core.cells import numpy_to_idarr
+from pyvista.utilities.cells import numpy_to_idarr
 from pyvista.core.errors import NotAllTrianglesError
